requirements.txt
datasets==2.20.0
evaluate==0.4.2
matplotlib==3.8.3
nltk==3.8.2
numpy==1.26.4
onnx==1.13.1
onnxruntime==1.15.1
optimum==1.13.2
Pillow==10.4.0
scikit_learn==1.5.1
torch==2.2.1
tqdm==4.66.5
transformers[torch]==4.33.2
rouge==1.0.1
rouge_score==0.1.2
absl-py==2.1.0
jupyter
